refactor(#2438): move connectors/ → backends/connectors/ per LEGO tiers
Relocate connector framework to its correct LEGO tier (Backend pillar).
Connectors are backend implementations and belong under backends/.

Updated all imports in source, tests, and scripts. Also fix two
pre-existing type-ignore comments exposed by the move:
- base.py: replace type-ignore with cast(BaseModel, data)
- mount_hooks.py: replace type-ignore with cast("SkillDocMixin", backend)
