+++ src/nexus/backends/connectors/gmail/errors.py
@@ -6,8 +6,8 @@ Domain-specific errors are defined here.
 
 from __future__ import annotations
 
-from nexus.connectors.base import ErrorDef
-from nexus.connectors.base_errors import CHECKPOINT_ERRORS, TRAIT_ERRORS
+from nexus.backends.connectors.base import ErrorDef
+from nexus.backends.connectors.base_errors import CHECKPOINT_ERRORS, TRAIT_ERRORS
 
 # Domain-specific errors for Gmail operations
 _DOMAIN_ERRORS: dict[str, ErrorDef] = {
