+++ src/nexus/backends/connectors/hn/SKILL.md
