+++ src/nexus/backends/gcalendar_connector.py
@@ -39,9 +39,7 @@ import yaml
 
 from nexus.backends.backend import Backend
 from nexus.backends.cache_mixin import IMMUTABLE_VERSION, CacheConnectorMixin
-from nexus.backends.oauth_mixin import OAuthConnectorMixin
-from nexus.backends.registry import ArgType, ConnectionArg, register_connector
-from nexus.connectors.base import (
+from nexus.backends.connectors.base import (
     CheckpointMixin,
     ConfirmLevel,
     OpTraits,
@@ -51,12 +49,14 @@ from nexus.connectors.base import (
     ValidatedMixin,
     ValidationError,
 )
-from nexus.connectors.calendar.errors import ERROR_REGISTRY
-from nexus.connectors.calendar.schemas import (
+from nexus.backends.connectors.calendar.errors import ERROR_REGISTRY
+from nexus.backends.connectors.calendar.schemas import (
     CreateEventSchema,
     DeleteEventSchema,
     UpdateEventSchema,
 )
+from nexus.backends.oauth_mixin import OAuthConnectorMixin
+from nexus.backends.registry import ArgType, ConnectionArg, register_connector
 from nexus.core.exceptions import BackendError, NexusFileNotFoundError
 from nexus.core.response import HandlerResponse, timed_response
