+++ src/nexus/backends/hn_connector.py
@@ -45,8 +45,8 @@ import httpx
 
 from nexus.backends.backend import Backend
 from nexus.backends.cache_mixin import CacheConnectorMixin, SyncResult
+from nexus.backends.connectors.base import SkillDocMixin
 from nexus.backends.registry import ArgType, ConnectionArg, register_connector
-from nexus.connectors.base import SkillDocMixin
 from nexus.core.exceptions import BackendError, NexusFileNotFoundError
 from nexus.core.response import HandlerResponse, timed_response
 
@@ -182,7 +182,7 @@ class HNConnectorBackend(Backend, CacheConnectorMixin, SkillDocMixin):
 
         try:
             content = (
-                resources.files("nexus.connectors.hn")
+                resources.files("nexus.backends.connectors.hn")
                 .joinpath("SKILL.md")
                 .read_text(encoding="utf-8")
             )
