+++ tests/e2e/self_contained/test_gcalendar_connector.py
@@ -19,14 +19,14 @@ from unittest.mock import MagicMock, patch
 import pytest
 from pydantic import ValidationError as PydanticValidationError
 
-from nexus.backends.local import LocalBackend
-from nexus.connectors.base import ValidationError
-from nexus.connectors.calendar.schemas import (
+from nexus.backends.connectors.base import ValidationError
+from nexus.backends.connectors.calendar.schemas import (
     CreateEventSchema,
     DeleteEventSchema,
     TimeSlot,
     UpdateEventSchema,
 )
+from nexus.backends.local import LocalBackend
 from nexus.core.config import PermissionConfig
 from nexus.core.permissions import OperationContext
 from nexus.factory import create_nexus_fs
