+++ tests/e2e/self_contained/test_gmail_connector.py
@@ -19,13 +19,13 @@ from unittest.mock import MagicMock, patch
 import pytest
 from pydantic import ValidationError as PydanticValidationError
 
-from nexus.backends.local import LocalBackend
-from nexus.connectors.base import ValidationError
-from nexus.connectors.gmail.schemas import (
+from nexus.backends.connectors.base import ValidationError
+from nexus.backends.connectors.gmail.schemas import (
     DraftEmailSchema,
     ReplyEmailSchema,
     SendEmailSchema,
 )
+from nexus.backends.local import LocalBackend
 from nexus.core.config import PermissionConfig
 from nexus.core.permissions import OperationContext
 from nexus.factory import create_nexus_fs
@@ -441,7 +441,7 @@ class TestOperationTraits:
 
     def test_send_email_traits(self, gmail_backend):
         """Test send_email has correct traits."""
-        from nexus.connectors.base import ConfirmLevel, Reversibility
+        from nexus.backends.connectors.base import ConfirmLevel, Reversibility
 
         traits = gmail_backend.get_operation_traits("send_email")
 
@@ -452,7 +452,7 @@ class TestOperationTraits:
 
     def test_reply_email_traits(self, gmail_backend):
         """Test reply_email has correct traits."""
-        from nexus.connectors.base import ConfirmLevel, Reversibility
+        from nexus.backends.connectors.base import ConfirmLevel, Reversibility
 
         traits = gmail_backend.get_operation_traits("reply_email")
 
@@ -462,7 +462,7 @@ class TestOperationTraits:
 
     def test_forward_email_traits(self, gmail_backend):
         """Test forward_email has correct traits."""
-        from nexus.connectors.base import ConfirmLevel, Reversibility
+        from nexus.backends.connectors.base import ConfirmLevel, Reversibility
 
         traits = gmail_backend.get_operation_traits("forward_email")
 
@@ -472,7 +472,7 @@ class TestOperationTraits:
 
     def test_draft_traits(self, gmail_backend):
         """Test create_draft has correct traits."""
-        from nexus.connectors.base import ConfirmLevel, Reversibility
+        from nexus.backends.connectors.base import ConfirmLevel, Reversibility
 
         traits = gmail_backend.get_operation_traits("create_draft")
