+++ tests/unit/connectors/test_checkpoint_mixin.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from nexus.connectors.base import CheckpointMixin, OpTraits, Reversibility
+from nexus.backends.connectors.base import CheckpointMixin, OpTraits, Reversibility
 
 
 class FakeCheckpointConnector(CheckpointMixin):
